Move COM to Win32 Lib Initialize COM once for all of pyglet.
