+++ pyglet/libs/win32/__init__.py
@@ -33,12 +33,13 @@
 # POSSIBILITY OF SUCH DAMAGE.
 # ----------------------------------------------------------------------------
 
+import atexit
 import struct
 
 import pyglet
+from . import com
 from . import constants
 from .types import *
-from pyglet import com
 
 IS64 = struct.calcsize("P") == 8
 
@@ -46,12 +47,14 @@ _debug_win32 = pyglet.options['debug_win32']
 
 if _debug_win32:
     import traceback
+
     _GetLastError = windll.kernel32.GetLastError
     _SetLastError = windll.kernel32.SetLastError
     _FormatMessageA = windll.kernel32.FormatMessageA
 
     _log_win32 = open('debug_win32.log', 'w')
-    
+
+
     def format_error(err):
         msg = create_string_buffer(256)
         _FormatMessageA(constants.FORMAT_MESSAGE_FROM_SYSTEM,
@@ -62,7 +65,8 @@ if _debug_win32:
                         len(msg),
                         c_void_p())
         return msg.value
-    
+
+
     class DebugLibrary:
         def __init__(self, lib):
             self.lib = lib
@@ -83,7 +87,6 @@ if _debug_win32:
             return f
 else:
     DebugLibrary = lambda lib: lib
-
 
 _gdi32 = DebugLibrary(windll.gdi32)
 _kernel32 = DebugLibrary(windll.kernel32)
@@ -122,7 +125,7 @@ _gdi32.GetCharABCWidthsW.restype = BOOL
 _gdi32.GetCharABCWidthsW.argtypes = [HDC, UINT, UINT, POINTER(ABC)]
 _gdi32.GetCharWidth32W.restype = BOOL
 _gdi32.GetCharWidth32W.argtypes = [HDC, UINT, UINT, POINTER(INT)]
-_gdi32.GetStockObject.restype =  HGDIOBJ
+_gdi32.GetStockObject.restype = HGDIOBJ
 _gdi32.GetStockObject.argtypes = [c_int]
 _gdi32.GetTextMetricsA.restype = BOOL
 _gdi32.GetTextMetricsA.argtypes = [HDC, POINTER(TEXTMETRIC)]
@@ -173,7 +176,8 @@ _user32.ClipCursor.argtypes = [LPRECT]
 _user32.CreateIconIndirect.restype = HICON
 _user32.CreateIconIndirect.argtypes = [POINTER(ICONINFO)]
 _user32.CreateWindowExW.restype = HWND
-_user32.CreateWindowExW.argtypes = [DWORD, c_wchar_p, c_wchar_p, DWORD, c_int, c_int, c_int, c_int, HWND, HMENU, HINSTANCE, LPVOID]
+_user32.CreateWindowExW.argtypes = [DWORD, c_wchar_p, c_wchar_p, DWORD, c_int, c_int, c_int, c_int, HWND, HMENU,
+                                    HINSTANCE, LPVOID]
 _user32.DefWindowProcW.restype = LRESULT
 _user32.DefWindowProcW.argtypes = [HWND, UINT, WPARAM, LPARAM]
 _user32.DestroyWindow.restype = BOOL
@@ -191,8 +195,8 @@ _user32.GetClientRect.argtypes = [HWND, LPRECT]
 _user32.GetCursorPos.restype = BOOL
 _user32.GetCursorPos.argtypes = [LPPOINT]
 # workaround for win 64-bit, see issue #664
-_user32.GetDC.restype = c_void_p # HDC
-_user32.GetDC.argtypes = [c_void_p] # [HWND]
+_user32.GetDC.restype = c_void_p  # HDC
+_user32.GetDC.argtypes = [c_void_p]  # [HWND]
 _user32.GetDesktopWindow.restype = HWND
 _user32.GetDesktopWindow.argtypes = []
 _user32.GetKeyState.restype = c_short
@@ -226,8 +230,8 @@ _user32.RegisterHotKey.argtypes = [HWND, c_int, UINT, UINT]
 _user32.ReleaseCapture.restype = BOOL
 _user32.ReleaseCapture.argtypes = []
 # workaround for win 64-bit, see issue #664
-_user32.ReleaseDC.restype = c_int32 # c_int
-_user32.ReleaseDC.argtypes = [c_void_p, c_void_p] # [HWND, HDC]
+_user32.ReleaseDC.restype = c_int32  # c_int
+_user32.ReleaseDC.argtypes = [c_void_p, c_void_p]  # [HWND, HDC]
 _user32.ScreenToClient.restype = BOOL
 _user32.ScreenToClient.argtypes = [HWND, LPPOINT]
 _user32.SetCapture.restype = HWND
@@ -275,13 +279,13 @@ _user32.GetRawInputData.argtypes = [HRAWINPUT, UINT, LPVOID, PUINT, UINT]
 _user32.ChangeWindowMessageFilterEx.restype = BOOL
 _user32.ChangeWindowMessageFilterEx.argtypes = [HWND, UINT, DWORD, c_void_p]
 
-#dwmapi
+# dwmapi
 _dwmapi.DwmIsCompositionEnabled.restype = c_int
 _dwmapi.DwmIsCompositionEnabled.argtypes = [POINTER(INT)]
 _dwmapi.DwmFlush.restype = c_int
 _dwmapi.DwmFlush.argtypes = []
 
-#_shell32
+# _shell32
 _shell32.DragAcceptFiles.restype = c_void
 _shell32.DragAcceptFiles.argtypes = [HWND, BOOL]
 _shell32.DragFinish.restype = c_void
@@ -306,9 +310,23 @@ _ole32.CoCreateInstance.argtypes = [com.REFIID, c_void_p, DWORD, com.REFIID, c_v
 _ole32.CoSetProxyBlanket.restype = HRESULT
 _ole32.CoSetProxyBlanket.argtypes = (c_void_p, DWORD, DWORD, c_void_p, DWORD, DWORD, c_void_p, DWORD)
 
-
 # oleaut32
 _oleaut32.VariantInit.restype = c_void_p
 _oleaut32.VariantInit.argtypes = [c_void_p]
 _oleaut32.VariantClear.restype = HRESULT
 _oleaut32.VariantClear.argtypes = [c_void_p]
+
+# Initialize COM in MTA mode. Required for: WIC (DirectWrite), WMF, and XInput
+try:
+    _ole32.CoInitializeEx(None, constants.COINIT_MULTITHREADED)
+except OSError as err:
+    warnings.warn("Could not set COM MTA mode. Unexpected behavior may occur.")
+
+
+def _uninitialize():
+    try:
+        _ole32.CoUninitialize()
+    except OSError:
+        pass
+
+atexit.register(_uninitialize)
