+++ pyglet/libs/win32/com.py
